+++ package.json
@@ -16,7 +16,7 @@
     "clean": "rimraf dist-es5 dist-es2015",
     "dc": "npm run dependency-check",
     "dependency-check": "dependency-check . --unused --no-dev && dependency-check . --missing --no-dev",
-    "lint": "tslint \"src/**/*.ts\"",
+    "lint": "tslint -p tsconfig.json --type-check \"src/**/*.ts\"",
     "prepare": "typings i && npm run build",
     "postpublish": "git push",
     "preversion": "git pull && npm run verify",
