+++ src/Cli.setLogLevel.spec.ts
@@ -1,20 +1,20 @@
 import { logLevel, getLevel } from 'aurelia-logging'
 import test from 'ava'
 
-import { createNoOpCli } from './test/commands'
+import { createNoOpCli, createArgv } from './test/util'
 
 const cli = createNoOpCli()
 
 test('with -V', t => {
-  cli.run(['node', 'cli', 'noop', '-V'])
+  cli.run(createArgv('noop', '-V'))
   t.is(getLevel(), logLevel.debug)
 })
 test('with --verbose', t => {
-  cli.run(['node', 'cli', 'noop', '--verbose'])
+  cli.run(createArgv('noop', '--verbose'))
   t.is(getLevel(), logLevel.debug)
 })
 test('with --silent', t => {
-  cli.run(['node', 'cli', 'noop', '--silent'])
+  cli.run(createArgv('noop', '--silent'))
   t.is(getLevel(), logLevel.none)
 })
