+++ src/Cli.showhelp.spec.ts
@@ -1,7 +1,7 @@
 import test from 'ava'
 import { stub, SinonStub } from 'sinon'
 
-import { createNoCommandCli } from './test/commands'
+import { createNoCommandCli, createArgv } from './test/util'
 
 const cli = createNoCommandCli()
 let showHelp: SinonStub
@@ -12,21 +12,21 @@ test.afterEach(() => {
   showHelp.restore()
 })
 test('when called with no parameter', t => {
-  cli.run(['node', 'cli'])
+  cli.run(createArgv())
   t.true(showHelp.called)
 })
 test('when called with -h', t => {
-  cli.run(['node', 'cli', '-h'])
+  cli.run(createArgv('-h'))
 
   t.true(showHelp.called)
 })
 test('when called with --help', t => {
-  cli.run(['node', 'cli', '--help'])
+  cli.run(createArgv('--help'))
 
   t.true(showHelp.called)
 })
 test('when command not found', t => {
-  cli.run(['node', 'cli', 'some'])
+  cli.run(createArgv('some'))
 
   t.true(showHelp.called)
 })
